Initial version
Initial code for an app that the user can enter carbs consumed. The app
then sends the data to a python script which would be located on the
device that is to be connected to the Carelink USB. This version is not
fully working and is definitely a rough draft.